refactor(arch): move _wrapper.py to bytecode/ with re-export shim (stage 3 of #569)
Third stage of the multi-stage package merge tracked in #569. Moves
the canonical home of RustWrapper / KirinRustWrapper from
bloqade.lanes._wrapper to bloqade.lanes.bytecode._wrapper — wrapper
machinery lives alongside the Rust types it wraps.

Old import path bloqade.lanes._wrapper remains as a thin re-export
shim, so every existing caller across bloqade.lanes, bloqade.gemini,
demos, and tests compiles unchanged. The shim is removed in the
final cleanup stage (Stage 9 of the plan) once all in-flight branches
have rebased onto the canonical path.

Refs #569.
